move struct44c8 into a proper location and name the file GameResources
